Made identifiers and type annot bases lists of strings.
The identifiers and type annotations were strings, but it's possible
that they are from different modules and therefore require `::` to
access. This was not possible for type annotations, while expressions
used a binaray operator to represent this access. Now both a type
annotation base and an identifier expression are lists of strings, and
the name access binary operator for expressions was removed.